requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
pydantic==2.5.0
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
sentence-transformers>=2.2.2
faiss-cpu>=1.8.0
torch>=2.2.0
transformers>=4.35.0
scikit-learn>=1.3.2
pandas>=2.1.3
numpy>=1.26.2
arxiv==2.1.0
pytest>=7.4.3
pytest-cov>=4.1.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.1
